refactoring package, while they are in `me.philippheuer.twitch4j.tmi` package
